Bugfixes, more features, improved performance
This is a bit more robust than previous versions, but is also a bit
larger (in terms of file size). I think the tradeoff is worth it, as the
new features are necessary when using the draggable in any but the
simplest applications.

Note that the js files should be included at the *end of the body* in
order to avoid onDomReady sniffing. I've also removed the AMD
integration because I'm not familiar with the syntax, could you add it
back? I'm also not familiar with the unit testing framework you're
using, but I have tested the script manually in all supported browsers.